requirements.txt
google-cloud-bigquery==2.6.1
google-cloud-bigquery-storage==2.1.0
google-cloud-storage==1.36.1
pandas==1.1.5
pyarrow==2.0.0
scipy==1.5.4
fire==0.4.0
